backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
llvmlite==0.40.1
numba==0.57.1
numpy==1.24.4
pandas==2.0.3
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
